MPM-1094
Added try/catch to SmtpClient.send().  Refactored web.config settings to
allow setting multiple smtp servers in the web.confing file instead of
using hard coded strings in the service.  EmailMessenger.send now
returns a boolean value indicated whether or not it was able to
successfully connect and send to the smpt server.  If any errors happen,
they are logged to the database using nLog.